+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # 0.3.0
 
-## Breaking Changes
+## Features
 
-
+- allow defaults to be a function on api
 
 # 0.2.5